+++ packages/marvel/src/Http/Resources/FaqResource.php
@@ -18,6 +18,8 @@ class FaqResource extends Resource
             'id'              => $this->id,
             'faq_title'       => request()->routeIs('faqs.index') ? $this->getTranslation('faq_title', app()->getLocale()) : $this->getRawOriginal('faq_title'),
             'faq_description' => request()->routeIs('faqs.index') ? $this->getTranslation('faq_description', app()->getLocale()) : $this->getRawOriginal('faq_description'),
+            'status'          => (int) $this->status,
+            'order'           => (int) $this->order,
         ];
     }
 }